Moves acknowledger logic to support package
